renamed blogData to Data
